+++ src/test/java/roomescape/dto/InputValidatorTest.java
@@ -3,7 +3,7 @@ package roomescape.dto;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import roomescape.dto.exception.InputNotAllowedException;
+import roomescape.dto.request.exception.InputNotAllowedException;
 
 class InputValidatorTest {
